requirements.txt
numpy
flask
flask-cors
requests<2.32
paho-mqtt
prometheus_client
flwr==1.11.1
typer>=0.9,<0.10
datetime
statsmodels
tensorflow
tensorflow_datasets
flwr.datasets
google==2.0.3
pandas>=1.4.0,!=1.5.0,!=2.0.0
scikit-learn>=1.3.0
kaleido
scikit-learn-extra
python-dotenv